[Event Hubs Client] Restructure Reading from a Single Partition (#13155)
The focus of these changes is to restructure the approach used to read events
from a single partition.  Previously, the background publishing mechanism was
used to try to ensure a smooth streaming of events to the iterator at the cost
of a longer pipeline and potentially higher memory use.  This was needed, in
large part, because no options were exposed for influencing the prefetch buffer
and batch size.

With the prefetch count and batch size now configurable, developers are able
to better tune the performance of reading to suit their specific application
scenarios, allowing a more efficient and direct use of a transport consumer to
read events and obviating the need for background publishing.